Separate configuration hooks
